+Various changes and additions
Actually implemented the second stage. It's a work in progress, obviously, but it's fairly close to done since I used stage 1 as a base :). I adjusted some existing attacks in stage 1 and made them harder to dodge, so that the difficulty remains more consistent to stage 2. Some of Owl girl's attacks now rotate, because I now just found out that's a built-in feature of BulletHellper